requirements.txt
python-arango
fastapi
validators
pytube
paramiko
scp
scenedetect==0.6.0.3
git+https://github.com/openai/CLIP.git
timm
wget
fairscale
openai==0.25.0
langchain
accelerate
spacy_wordnet
fastapi
facenet-pytorch==2.5.2
hdbscan
weaviate-client